% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_package.R
\name{inspect_package}
\alias{inspect_package}
\title{Inspect packages for structure and file types}
\usage{
inspect_package(path = ".")
}
\arguments{
\item{path}{A character vector or list of local package paths.}
}
\value{
A \code{pkginspectr_tbl} with one row per package containing structural
summaries.
}
\description{
Examine the file structure of one or more local packages and return summary
statistics including file counts, whether files contain functions and the
package version.
}
\examples{
\dontrun{
# Inspect the installed "stats" package
inspect_package(system.file(package = "stats"))
}
}
